+++ aten/src/ATen/native/cpu/RangeFactoriesKernel.cpp
@@ -1,6 +1,7 @@
 #include <cmath>
 #include <ATen/Config.h>
 #include <ATen/Dispatch.h>
+#include <ATen/native/DispatchStub.h>
 
 #include <ATen/AccumulateType.h>
 #include <ATen/cpu/vec256/vec256.h>
