+++ aten/src/ATen/native/quantized/cpu/int_repr_quant.cpp
@@ -1,5 +1,7 @@
+#include <ATen/ATen.h>
 #include <ATen/native/TensorIterator.h>
 #include <ATen/native/cpu/Loops.h>
+#include <ATen/native/DispatchStub.h>
 
 namespace at {
 namespace native {
