+++ aten/src/ATen/native/quantized/cpu/make_per_tensor_quantized_tensor.cpp
@@ -1,5 +1,7 @@
 #include <ATen/native/TensorIterator.h>
 #include <ATen/native/cpu/Loops.h>
+#include <ATen/Dispatch.h>
+#include <ATen/Functions.h>
 
 namespace at {
 namespace native {
