+++ caffe2/CMakeLists.txt
@@ -1319,8 +1319,18 @@ if(BUILD_TEST)
         list(GET CPU_CAPABILITY_NAMES ${i} CPU_CAPABILITY)
         list(GET CPU_CAPABILITY_FLAGS ${i} FLAGS)
         separate_arguments(FLAGS UNIX_COMMAND "${FLAGS}")
-        add_executable(${test_name}_${CPU_CAPABILITY} "${test_src}")
-        target_link_libraries(${test_name}_${CPU_CAPABILITY} torch_library gtest_main)
+        # Build vec256 with minimal dependencies on all platforms but Windows
+        if(NOT MSVC)
+          add_executable(${test_name}_${CPU_CAPABILITY} "${test_src}" ../aten/src/ATen/native/quantized/affine_quantizer_base.cpp)
+          # TODO: Get rid of c10 dependency (which is only needed for the implementation of AT_ERROR)
+          target_link_libraries(${test_name}_${CPU_CAPABILITY} c10 sleef gtest_main)
+          if(USE_FBGEMM)
+            target_link_libraries(${test_name}_${CPU_CAPABILITY} fbgemm)
+          endif()
+        else()
+          add_executable(${test_name}_${CPU_CAPABILITY} "${test_src}")
+          target_link_libraries(${test_name}_${CPU_CAPABILITY} torch_library gtest_main)
+        endif()
         target_include_directories(${test_name}_${CPU_CAPABILITY} PRIVATE $<INSTALL_INTERFACE:include>)
         target_include_directories(${test_name}_${CPU_CAPABILITY} PRIVATE $<BUILD_INTERFACE:${CMAKE_BINARY_DIR}/include>)
         target_include_directories(${test_name}_${CPU_CAPABILITY} PRIVATE ${ATen_CPU_INCLUDE})
